refactored page.tsx into the Game component so that the main part of the game which tracks dollar amount can be clientsided
